# optimization: speed up block serialization