@l0rinc 2025-04-17
@TheCharlatan, I've added back the single-byte optimizations (simplified slightly after rebase), your feedback is appreciated.